Lazy load package, update config, remove extra throttling
